requirements.txt
asgiref==3.4.1
autopep8==1.6.0
boto3==1.19.4
botocore==1.22.4
certifi==2021.5.30
colorama==0.4.4
Cython==0.29.23
Django==3.2
django-storages==1.12.2
djangorestframework==3.12.4
gensim==4.1.2
implicit==0.4.8
jmespath==0.10.0
mysqlclient==2.0.3
numpy==1.21.2
pandas==1.3.3
pycodestyle==2.8.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
scipy==1.7.1
six==1.16.0
smart-open==5.2.1
sqlparse==0.4.2
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
wincertstore==0.2